backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Tomasz Marchlewski"]

[tool.poetry.dependencies]
python = "^3.12.2"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
llama-cpp-python = "^0.2.57"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
app = "src.app.main:run"
